usize -> isize cast

+++ src/raw_vec.rs
@@ -72,11 +72,11 @@ impl<T> RawVec<T>
 
 impl<T> Index<usize> for RawVec<T> {
     type Output = T;
-    fn index(&self, index: usize) -> &Self::Output { unsafe { &*self.data.offset(index) } }
+    fn index(&self, index: usize) -> &Self::Output { unsafe { &*self.data.offset(index as isize) } }
 }
 
 impl<T> IndexMut<usize> for RawVec<T> {
     fn index_mut(&mut self, index: usize) -> &mut Self::Output {
-        unsafe { &mut *self.data.offset(index) }
+        unsafe { &mut *self.data.offset(index as isize) }
     }
 }
